Check if patient with COVID-related AE leading to STUDY discon and also has STUDY Discon record in DS

Flag if patient has a record with COVID-related AE where AE.AEDECOD matches covid.REFTERM and leads to STUDY discon (AE.AEACNOT*=SUBJECT DISCONTINUED FROM STUDY) but has no STUDY Discon record in DS (DS.DSSCAT=STUDY DISCONTINUATION)

check_ae_aeacnoth_ds_stddisc_covid(AE, DS, covid_df = NULL)

Arguments

AE
Adverse Events SDTM dataset with variables USUBJID,AEDECOD,AEACNOT* (can be multiple vars)

DS
Discon SDTM dataset with variables USUBJID,DSSCAT,DSDECOD

covid_df
Dataframe of AE terms identify covid, contains variable REFTERM

Value

boolean value if check failed or passed with 'msg' attribute if the test failed

Author

Natalie Springfield

Examples

covid_df = data.frame(REFTERM = c("COVID-19",
                                  "CORONAVIRUS POSITIVE"
                                  )
                     )

AE <- data.frame(
 USUBJID = 1:5,
 AEDECOD = c("covid-19", "covid-19", "covid-19","Some AE", "CORONAVIRUS POSITIVE" ),
 AEACNOTH=c("SUBJECT DISCONTINUED FROM STUDY",
            "NONE",
            "NONE", 
            "SUBJECT DISCONTINUED FROM STUDY",
            "NONE"),
 AEACNOTH1=c("SUBJECT DISCONTINUED FROM STUDY", 
             "NONE",
             "SUBJECT DISCONTINUED FROM STUDY",
             "NONE", 
             "SUBJECT DISCONTINUED FROM STUDY"),
 AEACNOTH2=c("SUBJECT DISCONTINUED FROM STUDY", 
             "NONE",
             "NONE",
             "SUBJECT DISCONTINUED FROM STUDY",
             "NONE")
)

DS <- data.frame(
 USUBJID = 1:3,
 DSSCAT=c("TREATMENT DISCONTINUATION","STUDY DISCONTINUATION","STUDY DISCONTINUATION"),
 DSDECOD="DISCON REASON"
)

check_ae_aeacnoth_ds_stddisc_covid(AE,DS,covid_df = covid_df)

#> [1] FALSE
#> attr(,"msg")
#> [1] "Found 2 patient(s) with COVID-related AEs leading to Study Discon, but no corresponding Study Discon recs in DS. "
#> attr(,"data")
#>   USUBJID              AEDECOD DSDECOD DSSCAT                        AEACNOTH
#> 1       1             COVID-19    <NA>   <NA> SUBJECT DISCONTINUED FROM STUDY
#> 3       5 CORONAVIRUS POSITIVE    <NA>   <NA>                            NONE
#>                         AEACNOTH1                       AEACNOTH2
#> 1 SUBJECT DISCONTINUED FROM STUDY SUBJECT DISCONTINUED FROM STUDY
#> 3 SUBJECT DISCONTINUED FROM STUDY                            NONE