Check MedDRA version in TS and return dataframe with covid terms according to this version

This function reads in a dataset with covid terms according to MedDRA version found in TS This function assumes you have a folder called "data" that contains .Rdata files with file names indexed by medDRA version, e.g. covid230.Rdata, covid231.Rdata, covid240.Rdata,etc. It also assumes a pan MedDRA version data/covid.Rdata file which it will use if the specific version cant be found.

roche_covid_df(TS = NULL)

Arguments

TS
Trial Summary SDTM dataset (optional) with variables TSPARMCD, TSVAL

Value

dataframe

Examples

if (FALSE) {
#First COVID dataset is from v23.0, so use the latest
ts1 <- data.frame(TSPARMCD = "AEDICT", TSVAL = c("MedDRA 22.0"))
roche_covid_df(TS=ts1)

#COVID datset chosen regardless of the 'v' precending version number
ts2 <- data.frame(TSPARMCD = "AEDICT", TSVAL = c("MedDRA v23.0"))
roche_covid_df(TS=ts2)

ts2a <- data.frame(TSPARMCD = "AEDICT", TSVAL = c("MedDRA 23.0"))
roche_covid_df(TS=ts2a)

#Missing version default to use the latest available
ts3 <- data.frame(TSPARMCD = "AEDICT", TSVAL = "")
roche_covid_df(TS=ts3)

#Future version default to use the latest available
ts4 <- data.frame(TSPARMCD = "AEDICT", TSVAL = c("MedDRA v40.1"))
roche_covid_df(TS=ts4)
}