Check for consistency between new lesions and overall PD response

This checks for patients with new lesions in TU (TUSTRESC=='NEW') but no overall response of PD (RSTESTCD == 'OVRLRESP' and RSSTRESC = 'PD') in RS. Only applies to assessments by investigator, if TUEVAL and RSEVAL variables available.

check_tu_rs_new_lesions(RS, TU)

Arguments

RS

Response SDTM dataset with variables USUBJID, RSSTRESC, RSTESTCD

TU

Tumor Identification SDTM dataset with variables USUBJID, TUSTRESC, TUDTC

Value

TRUE if check is passed and FALSE if check failed + 'msg' and 'data' attributes

Author

Will Harris

Examples

TU <- data.frame(
 USUBJID = 1:3,
 TUSTRESC = c("INV001","NEW","NEW"),
 TUDTC = "2017-01-01"
)

RS <- data.frame(
 USUBJID = 1:2,
 RSSTRESC = c("SD","NE")
)

# required variable is missing 
check_tu_rs_new_lesions(RS,TU)
#> [1] FALSE
#> attr(,"msg")
#> [1] "RS is missing the variable: RSTESTCD"

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
RS <- RS %>%
  mutate(RSTESTCD = 'OVRLRESP')

# flag USUBJIDs with NEW 
check_tu_rs_new_lesions(RS,TU)
#> [1] FALSE
#> attr(,"msg")
#> [1] "TU has 2 patient(s) with a new lesion but no overall response of PD. "
#> attr(,"data")
#>   USUBJID      TUDTC
#> 1       2 2017-01-01
#> 2       3 2017-01-01

RS$RSSTRESC[2] = "PD"

# flag USUBJID with NEW and without PD
check_tu_rs_new_lesions(RS,TU)
#> [1] FALSE
#> attr(,"msg")
#> [1] "TU has 1 patient(s) with a new lesion but no overall response of PD. "
#> attr(,"data")
#>   USUBJID      TUDTC
#> 1       3 2017-01-01

TU <- TU %>%
mutate(TUEVAL = "INDEPENDENT ASSESSOR")

RS <- RS %>% 
mutate(RSEVAL = "INDEPENDENT ASSESSOR")

## pass if by IRF, even if NEW in TU
check_tu_rs_new_lesions(RS,TU)
#> [1] TRUE

RS <- NULL

# required dataset missing 
check_tu_rs_new_lesions(RS,TU)
#> [1] FALSE
#> attr(,"msg")
#> [1] "RS is missing the variables: USUBJID, RSSTRESC, RSTESTCD"